+++ monoloco/prep/preprocess_ki.py
@@ -8,7 +8,7 @@ from collections import defaultdict
 import json
 import datetime
 
-from ..features.transforms import transform_keypoints
+from ..prep.transforms import transform_keypoints
 from ..utils.kitti import get_calibration, split_training, parse_ground_truth
 from ..utils.monoloco import get_monoloco_inputs
 from ..utils.pifpaf import preprocess_pif
