+++ tests/test_utils.py
@@ -1,10 +1,7 @@
 
 
-from utils.iou import get_iou_matrix
-from utils.camera import pixel_to_camera
-
-
 def test_iou():
+    from monoloco.utils.iou import get_iou_matrix
     boxes_pred = [[1, 100, 1, 200]]
     boxes_gt = [[100., 120., 150., 160.],[12, 110, 130., 160.]]
     iou_matrix = get_iou_matrix(boxes_pred, boxes_gt)
@@ -12,6 +9,7 @@ def test_iou():
 
 
 def test_pixel_to_camera():
+    from monoloco.utils.camera import pixel_to_camera
     kk = [[718.3351, 0., 600.3891], [0., 718.3351, 181.5122], [0., 0., 1.]]
     zz = 10
     uv_vector = [1000., 400.]
